requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.6.0; python_version >= '3.9'
bcrypt==4.2.0; python_version >= '3.7'
certifi==2024.8.30; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
dnspython==2.7.0; python_version >= '3.9'
ecdsa==0.19.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
email-validator==2.2.0; python_version >= '3.8'
fastapi[all]==0.115.0; python_version >= '3.8'
fastapi-cache2[redis]==0.2.2; python_version >= '3.8' and python_version < '4.0'
fastapi-cli[standard]==0.0.5; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.6; python_version >= '3.8'
httptools==0.6.1; python_full_version >= '3.8.0'
httpx==0.27.2; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
iniconfig==2.0.0; python_version >= '3.7'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.1; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
orjson==3.10.7; python_version >= '3.8'
packaging==24.1; python_version >= '3.8'
pendulum==3.0.0; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pydantic==2.9.2; python_version >= '3.8'
pydantic-core==2.23.4; python_version >= '3.8'
pydantic-extra-types==2.9.0; python_version >= '3.8'
pydantic-settings==2.5.2; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pytest==8.3.3; python_version >= '3.8'
pytest-asyncio==0.24.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
python-jose==3.3.0
python-multipart==0.0.12; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
redis==4.6.0; python_version >= '3.7'
rich==13.9.2; python_full_version >= '3.8.0'
rsa==4.9; python_version >= '3.6' and python_version < '4'
shellingham==1.5.4; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.38.6; python_version >= '3.8'
time-machine==2.16.0; python_version >= '3.9'
typer==0.12.5; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.2; python_version >= '2'
ujson==5.10.0; python_version >= '3.8'
uvicorn[standard]==0.31.0; python_version >= '3.8'
uvloop==0.20.0; python_full_version >= '3.8.0'
watchfiles==0.24.0; python_version >= '3.8'
websockets==13.1; python_version >= '3.8'
